ensure permissions can only be set by users with elevated access

+++ src/models/RootGroupPermission.ts
@@ -1,6 +1,7 @@
-import { Access, PrismaClient, RootGroupPermission as DbGroupPermission } from '@prisma/client';
+import { Access, PrismaClient, RootGroupPermission as DbGroupPermission, User, Role } from '@prisma/client';
 import prisma from '../prisma';
 import { asGroupAccess } from '../helpers/accessPolicy';
+import { hasElevatedAccess } from './User';
 
 // TODO: Consider checking existence of documentRoot / studentGroup to provide better error messages / exceptions.
 
@@ -17,14 +18,6 @@ export type CompleteApiGroupPermission = {
     access: Access;
 };
 
-function asApiRecord(dbResult: DbGroupPermission): ApiGroupPermission {
-    return {
-        id: dbResult.id,
-        groupId: dbResult.studentGroupId,
-        access: dbResult.access
-    };
-}
-
 function asCompleteApiRecord(dbResult: DbGroupPermission): CompleteApiGroupPermission {
     return {
         id: dbResult.id,
@@ -33,14 +26,38 @@ function asCompleteApiRecord(dbResult: DbGroupPermission): CompleteApiGroupPermi
         access: dbResult.access
     };
 }
+const ensureAccessOrThrow = async (actor: User, groupId: string) => {
+    const group = await prisma.studentGroup.findUnique({
+        where: {
+            id: groupId,
+            users: {
+                some: {
+                    userId: actor.id,
+                    isAdmin: true
+                }
+            }
+        }
+    });
+    if (!group) {
+        throw new Error('Not authorized');
+    }
+};
 
 function RootGroupPermission(db: PrismaClient['rootGroupPermission']) {
     return Object.assign(db, {
         async createModel(
+            actor: User,
             documentRootId: string,
             studentGroupId: string,
             access: Access
         ): Promise<CompleteApiGroupPermission> {
+            if (!hasElevatedAccess(actor.role)) {
+                throw new Error('Not authorized');
+            }
+            if (actor.role === Role.TEACHER) {
+                ensureAccessOrThrow(actor, studentGroupId);
+            }
+
             const result = await db.create({
                 data: {
                     documentRootId: documentRootId,
@@ -51,7 +68,18 @@ function RootGroupPermission(db: PrismaClient['rootGroupPermission']) {
             return asCompleteApiRecord(result);
         },
 
-        async updateModel(id: string, access: Access): Promise<CompleteApiGroupPermission> {
+        async updateModel(actor: User, id: string, access: Access): Promise<CompleteApiGroupPermission> {
+            if (!hasElevatedAccess(actor.role)) {
+                throw new Error('Not authorized');
+            }
+            const record = await db.findUnique({ where: { id: id } });
+            if (!record) {
+                throw new Error('User permission not found');
+            }
+            if (actor.role === Role.TEACHER) {
+                ensureAccessOrThrow(actor, record.studentGroupId);
+            }
+
             const result = await db.update({
                 where: {
                     id: id
@@ -63,7 +91,18 @@ function RootGroupPermission(db: PrismaClient['rootGroupPermission']) {
             return asCompleteApiRecord(result);
         },
 
-        async deleteModel(id: string): Promise<CompleteApiGroupPermission> {
+        async deleteModel(actor: User, id: string): Promise<CompleteApiGroupPermission> {
+            if (!hasElevatedAccess(actor.role)) {
+                throw new Error('Not authorized');
+            }
+            const record = await db.findUnique({ where: { id: id } });
+            if (!record) {
+                throw new Error('User permission not found');
+            }
+            if (actor.role === Role.TEACHER) {
+                ensureAccessOrThrow(actor, record.studentGroupId);
+            }
+
             const result = await db.delete({
                 where: {
                     id: id
